package refactor ive been meaning to do for a while p2
